+++ packages/core/src/utils/index.ts
@@ -1,4 +1,3 @@
-export { hexValue, getAddress } from 'ethers/lib/utils'
-
 export { getInjectedName } from './getInjectedName'
 export { normalizeChainId } from './normalizeChainId'
+export { warn } from './warn'